backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
openai==1.10.0
Pillow==9.5.0
opencv-python==4.9.0.80
numpy==1.26.3
# Using the python port of material color utilities
material-color-utilities-python==0.1.5
requests
google-generativeai
sqlalchemy
supabase
